Add permissive option allowing for arbitrary Solr parameter
If the Solr config is set to permissive, all parameters are passed through to Solr.
If it is not permissive, issuing a Solr call with any non-vetted parameters will
result in an exception.

The default is false and this should only be enabled on local or devel for decoupling
ad hoc frontend tweaking of Solr calls from enabling of said tweaks in ds-discover.
The non-vetted arguments are logged on `warn` for later incorporation in ds-discover.